add usage example draw rectangle
