+++ public/usage-examples/graphics/draw_rectangle/draw_rectangle-1-staircase-pattern.txt
@@ -1,0 +1,3 @@
+#### Draw Rectangle
+
+The following code shows examples of using [Draw Rectangle](/api/graphics#draw-rectangle-1) to show how to draw a staircase pattern.
